Redcap integration
  1.  Allow meta-data for subject and project
  2.  Add an additional Authority for ROLE_EXTERNAL_ERF_INTEGRATOR with limited resource access
  3.  Create a UUID as a subject.login to uniquely identify a `Subject in the system
  4. Allow to discontinue a subject.
  5.  Additional API to query a subject based on projectId and externalId.

+++ src/main/java/org/radarcns/management/web/rest/SourceResource.java
@@ -153,6 +153,11 @@ public class SourceResource {
     @Secured({AuthoritiesConstants.SYS_ADMIN, AuthoritiesConstants.PROJECT_ADMIN})
     public ResponseEntity<Void> deleteSource(@PathVariable Long id) {
         log.debug("REST request to delete Source : {}", id);
+        if (sourceRepository.findOne(id).isAssigned()) {
+            return ResponseEntity.badRequest()
+                .headers(HeaderUtil.createFailureAlert(ENTITY_NAME, "sourceIsAssigned", "Cannot delete an assigned source"))
+                .body(null);
+        }
         sourceService.delete(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
     }
